+++ agagd/agagd_core/views/index.py
@@ -5,13 +5,8 @@ import datetime
 import agagd_core.models as agagd_models
 
 # Django Tables Imports
-from agagd_core.tables.beta import (
-    GamesTable,
-    PlayersTournamentTable,
-    Top10DanTable,
-    Top10KyuTable,
-    TournamentsTable,
-)
+from agagd_core.tables.beta import GamesTable, PlayersTournamentTable, TournamentsTable
+from agagd_core.tables.top_players import TopDanTable, TopKyuTable
 from django.shortcuts import render
 from django.template.response import TemplateResponse
 
@@ -72,13 +67,13 @@ class FrontPageView(DetailView):
         # Index Tables
         latest_games_table = GamesTable(self.__get_latest_games())
         latest_tournaments_table = TournamentsTable(self.__get_latest_tournaments())
-        top_10_dan_table = Top10DanTable(self.__get_top_dan())
-        top_10_kyu_table = Top10KyuTable(self.__get_top_kyu())
+        top_dan_table = TopDanTable(self.__get_top_dan())
+        top_kyu_table = TopKyuTable(self.__get_top_kyu())
 
         context = locals()
         context["latest_games_table"] = latest_games_table
         context["latest_tournaments_table"] = latest_tournaments_table
-        context["top_10_dan_table"] = top_10_dan_table
-        context["top_10_kyu_table"] = top_10_kyu_table
+        context["top_dan_table"] = top_dan_table
+        context["top_kyu_table"] = top_kyu_table
 
         return TemplateResponse(request, self.template_name, context)
