+++ agagd/jinja2/beta.index.html
@@ -2,21 +2,17 @@
 
 {% block content %}
     <section class="container">
-        <section class="row top-players-tables">
+        <section class="row">
             <section class="col-sm-12 col-md-12 col-lg-6 top-dan">
-                <h5 class="table-headers top-players-headers mb-0">Top Dan</h5>
+                <h5 class="table-headers mb-0">Top Dan</h5>
                 <p class="m-0 pb-1 small">Last twelve months</p>
-                <div class="top-dan-table">
-                    {{ top_10_dan_table.as_html(request) }}
-                </div>
+                {{ top_dan_table.as_html(request) }}
             </section>
 
-            <section class="col-sm-12 col-md-12 col-lg-6 top-kyu">
-                <h5 class="table-headers top-players-headers mb-0">Top Kyu</h5>
+            <section class="col-sm-12 col-md-12 col-lg-6">
+                <h5 class="table-headers mb-0">Top Kyu</h5>
                 <p class="m-0 pb-1 small">Last twelve months</p>
-                <div class="top-kyu-table">
-                    {{ top_10_kyu_table.as_html(request) }}
-                </div>
+                {{ top_kyu_table.as_html(request) }}
             </section>
         </section>
